Fixed error message, removed threadIds and intervalId from import history

+++ src/workers/utils.ts
@@ -3,47 +3,55 @@ import * as os from 'os';
 // tslint:disable-next-line
 import { Worker } from 'worker_threads';
 
-let workers: any[] = [];
+let workers: any[] = ['asd'];
+let intervals: any[] = [];
 
 export const createWorkers = (workerPath: string, workerData: any, results: string[]) => {
-  const threadIds: number[] = [];
-
   return new Promise((resolve, reject) => {
-    results.forEach(result => {
-      try {
-        const worker = new Worker(workerPath, {
-          workerData: {
-            ...workerData,
-            result,
-          },
-        });
+    if (workers.length > 0) {
+      return reject(new Error('Workers are busy'));
+    }
 
-        workers.push(worker);
-        threadIds.push(worker.threadId);
-
-        worker.on('message', () => {
-          workers = workers.filter(workerObj => {
-            return worker.threadId !== workerObj.threadId;
+    const interval = setImmediate(() => {
+      results.forEach(result => {
+        try {
+          const worker = new Worker(workerPath, {
+            workerData: {
+              ...workerData,
+              result,
+            },
           });
 
-          worker.terminate();
-        });
+          workers.push(worker);
 
-        worker.on('error', e => {
+          worker.on('message', () => {
+            workers = workers.filter(workerObj => {
+              return worker.threadId !== workerObj.threadId;
+            });
+
+            worker.terminate();
+          });
+
+          worker.on('error', e => {
+            reject(new Error(e));
+          });
+
+          worker.on('exit', code => {
+            if (code !== 0) {
+              reject(new Error(`Worker stopped with exit code ${code}`));
+            }
+          });
+        } catch (e) {
           reject(new Error(e));
-        });
+        }
+      });
 
-        worker.on('exit', code => {
-          if (code !== 0) {
-            reject(new Error(`Worker stopped with exit code ${code}`));
-          }
-        });
-      } catch (e) {
-        reject(new Error(e));
-      }
+      clearIntervals();
     });
 
-    resolve(threadIds);
+    intervals.push(interval);
+
+    resolve(true);
   });
 };
 
@@ -65,14 +73,18 @@ export const splitToCore = (datas: any[]) => {
   return results;
 };
 
-export const removeWorker = threadIds => {
-  workers = workers.filter(worker => {
-    const threadId = worker.threadId;
-
-    if (threadIds.includes(worker.threadId)) {
-      worker.terminate();
-    }
-
-    return !threadIds.includes(threadId);
+export const removeWorkers = () => {
+  workers.forEach(worker => {
+    worker.terminate();
   });
+
+  workers = [];
+};
+
+export const clearIntervals = () => {
+  intervals.forEach(interval => {
+    clearImmediate(interval);
+  });
+
+  intervals = [];
 };
